:hammer: Move Encoding types into Yarhl.Format
