+++ yarhl.Media/Text/Encodings/EscapeOutRangeEncoding.cs
@@ -23,7 +23,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.IO.Encodings
+namespace Yarhl.Media.Text.Encodings
 {
     using System;
     using System.Collections.Generic;
