+++ yarhl.UnitTests/Media/Text/Encodings/EscapeOutRangeEncodingTests.cs
@@ -23,13 +23,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.IO
+namespace Yarhl.UnitTests.Media.Text.Encodings
 {
     using System;
     using System.Linq;
     using System.Text;
     using NUnit.Framework;
-    using Yarhl.IO.Encodings;
+    using Yarhl.Media.Text.Encodings;
 
     [TestFixture]
     public class EscapeOutRangeEncodingTests
